requirements.txt
sqlparse==0.4.2
dj-database-url==1.0.0
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
drf-spectacular==0.26.2
gunicorn==20.1.0
jsonschema==4.14.0
inflection==0.5.1
Pillow==9.2.0
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.2.1
uritemplate==4.1.1
PyYAML==6.0
whitenoise==6.2.0
psycopg2-binary==2.9.6